requirements.txt
altgraph==0.17.2
cffi==1.15.0
macholib==1.16
Pillow==9.1.1
pycparser==2.21
pyheif==0.7.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
